requirements.txt
discord.py
python-dotenv
requests
google-generativeai
pytz